DEV: Fix theme compatibility problem
Why this change?

Technically `< 3.2.0.beta2-dev` is wrong cause it is a development
version and there are many commits within that version. If Discourse
core is on a commit that does not include the necessary changes required
in 865bb3308fbeacbd77dca793c2f60c2a3afbb464, the theme will break.