FIX: Make `links` and `icons` setting migration more resilient (#39)
This commit updates the migrations to assume that the old setting value
may be invalid so that we don't end up with a new setting which fails
the objects schema validation.

This is a follow up to 7e304e0a12115c6596fc64ff173368cba1a28e7b which
added migrations for the `links` and `icons` settings. We have
gotten reports of errors happening with the migrations because the value of the old
setting before migration did not follow the "informal" format required
previously. Since there was nothing enforcing the format for the value
previously, we need to assume the worst possible cases in our
migrations.